Test: Move non-test files to subfolders for client/lib/terms
